refactor(gallery): extract shared TemplateGallery, dynamic Others dedup, green verified badge; reuse in WelcomeDialog
