+++ packages/web/src/components/LibraryDialog.test.tsx
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import { render, screen, within } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import { LibraryDialog } from "./LibraryDialog";
 
 const TOP_INDEX = `# OKF Bundles
@@ -14,22 +14,12 @@ beforeEach(() => {
 });
 afterEach(() => vi.restoreAllMocks());
 
+// Detailed gallery/dedup behavior is covered by TemplateGallery.test.tsx; this
+// is a thin smoke test confirming the dialog chrome mounts the shared gallery.
 describe("LibraryDialog", () => {
-  it("renders a Verified Templates Gallery section from the fetched index", async () => {
+  it("renders the dialog header and mounts the template gallery", async () => {
     render(<LibraryDialog onUse={vi.fn()} onClose={vi.fn()} />);
+    expect(screen.getByText("Template library")).toBeInTheDocument();
     expect(await screen.findByText(/verified templates gallery/i)).toBeInTheDocument();
-    expect(await screen.findByText("E-Commerce")).toBeInTheDocument();
-    expect(await screen.findByText("SaaS")).toBeInTheDocument();
-  });
-
-  it("shows an Others section that excludes the deduped built-in templates", async () => {
-    render(<LibraryDialog onUse={vi.fn()} onClose={vi.fn()} />);
-    const others = await screen.findByTestId("others-section");
-    // Deduped built-ins must NOT appear in Others.
-    expect(within(others).queryByText("SaaS / Subscription")).not.toBeInTheDocument();
-    expect(within(others).queryByText("E-commerce / Retail")).not.toBeInTheDocument();
-    expect(within(others).queryByText("Finance / Fintech")).not.toBeInTheDocument();
-    // A non-deduped built-in still appears.
-    expect(within(others).getByText("Marketplace")).toBeInTheDocument();
   });
 });
